refactor(cli): deepen verified EQL bundle
